moved drop down menus to a component
