+++ client/src/components/nav/FooterNav.tsx
@@ -9,7 +9,7 @@ const FooterNav = ({ topics }: Props) => {
     <nav className="w-full text-white flex items-stretch">
       <ul className="w-full flex flex-wrap">
         {topics.map((topic) => (
-          <li className="w-1/6 border-r h-12 text-sm">
+          <li key={topic} className="w-1/6 border-r h-12 text-sm">
             <Link
               className={`h-full hover:bg-pink-800 flex items-center px-5`}
               to={`/${topic}`}
